Improvement: package loss field added to PingResponse
As requested on issue #85, in some cases We want to know How much packages has been lost on ping action. So, this pull request aims to add to PingResponse object a field called packageLoss Who holds this percentage.

Sometimes We want to set some tolerance for errors, but not forgeting that something isn't good. With that property you Could (for example) report that internet is unstable.

- fixed some mispelling at CONTRIBUTIONS.md
- fixed a broken link at CONTRIBUTIONS.md
- adjusted line length for testing pass at win.js file
- adjusted ping binary test scenario
- adjusted answer.json with the new property
- added two new fixtures with failling context (100% package failure and partial package failure)
- expected output at CONTRIBUTION.md file was updated
- updated parsed response explanation at README.md

Related: https://github.com/danielzzz/node-ping/issues/85